+++ library/polynomial/fps/formal_power_series.hpp
@@ -28,12 +28,12 @@ struct FormalPowerSeries : vector<T> {
     P operator%(const P &r) const { return P(*this) %= r; }
     P &operator+=(const P &r) {
         if (r.size() > this->size()) this->resize(r.size());
-        for (int i = 0; i < r.size(); i++) (*this)[i] += r[i];
+        for (int i = 0; i < (int)r.size(); i++) (*this)[i] += r[i];
         return *this;
     }
     P &operator-=(const P &r) {
         if (r.size() > this->size()) this->resize(r.size());
-        for (int i = 0; i < r.size(); i++) (*this)[i] -= r[i];
+        for (int i = 0; i < (int)r.size(); i++) (*this)[i] -= r[i];
         return *this;
     }
     // https://judge.yosupo.jp/problem/convolution_mod
@@ -61,7 +61,7 @@ struct FormalPowerSeries : vector<T> {
     }
     P operator-() const {
         P ret(this->size());
-        for (int i = 0; i < this->size(); i++) ret[i] = -(*this)[i];
+        for (int i = 0; i < (int)this->size(); i++) ret[i] = -(*this)[i];
         return ret;
     }
     P &operator+=(const T &r) {
@@ -75,12 +75,12 @@ struct FormalPowerSeries : vector<T> {
         return *this;
     }
     P &operator*=(const T &v) {
-        for (int i = 0; i < this->size(); i++) (*this)[i] *= v;
+        for (int i = 0; i < (int)this->size(); i++) (*this)[i] *= v;
         return *this;
     }
     P dot(P r) const {
         P ret(min(this->size(), r.size()));
-        for (int i = 0; i < ret.size(); i++) ret[i] = (*this)[i] * r[i];
+        for (int i = 0; i < (int)ret.size(); i++) ret[i] = (*this)[i] * r[i];
         return ret;
     }
     P operator>>(int sz) const {
